Directly use cpythons errors.c file
