+++ from_cpython/CMakeLists.txt
@@ -87,6 +87,7 @@ file(GLOB_RECURSE STDOBJECT_SRCS Objects
     complexobject.c
     dictobject.c
     dictproxy.c
+    errors.c
     exceptions.c
     floatobject.c
     fileobject.c
